+++ examples/bouncing_ball.py
@@ -68,7 +68,7 @@ block = BouncingBall(system, initial_velocity=[1, 0], initial_position=[0, 10])
 
 # Run a simulation for 10s
 simulator = Simulator(system, start_time=0.0)
-simulator.run_until(10.0)
+simulator.run_until(8.0)
 
 # Plot the x- and y-position of the ball against each other
 position = simulator.result.state[:, block.position.state_slice]